// src/types/normalizedPlayers.ts
export interface NormalizedPlayerResponse {
  id: string;
  playerId: null | string;
  rank?: number;
  playerName: string;
  teamAbr: string;
  pos: string;
  byeWeek: number | null;
  strengthOgSchedule?: string;
  avgAdp?: string;
  notes?: string;
  avgFanPoints?: number;
  avgPassingYds?: number;
  avgPassingTds?: number;
  avgReceivingRec?: number;
  avgReceivingYds?: number;
  avgReceivingTds?: number;
  avgRushingAtt?: number;
  avgRushingYds?: number;
  avgRushingTds?: number;
  totalFanPoints?: number;
  totalPassingYds?: number;
  totalPassingTds?: number;
  totalReceivingRec?: number;
  totalReceivingYds?: number;
  totalReceivingTds?: number;
  totalRushingAtt?: number;
  totalRushingYds?: number;
  totalRushingTds?: number;
  name: string;
  injuryStatus: string | null;
  depth: number | null;
  positionGroup: string | null;
  playerDepthPosition: string;
  playerImageSrc: string;
  height: string | null;
  weight: null | string;
  number: null | string;
  teamName?: string | null;
  teamImgSrc?: string | null;
  experience: string | null;
  drafted: boolean;
  getRoundAndPick?: string;
}